safari-600.7-branch/LayoutTests/fast/dom/Geolocation/notimer-after-unload-expected.txt

main frame - has 1 onunload handler(s)
ALERT: unload-called
You should have seen one unload alert appear.
